src/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
diagrams==0.23.3
graphviz==0.20.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
